merged: changes from trunk. revs: r24602:r24700
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/pvr-testing@24904 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ XBMC/xbmc/cores/dvdplayer/DVDCodecs/Video/DVDVideoCodec.h
@@ -42,7 +42,7 @@ struct DVDVideoPicture
 
   unsigned int iFlags;
 
-  unsigned int iRepeatPicture;
+  double       iRepeatPicture;
   double       iDuration;
   unsigned int iFrameType         : 4; // see defines above // 1->I, 2->P, 3->B, 0->Undef
   unsigned int color_matrix       : 4;
